# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fakegaming Bot API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login via /auth/login and paste the token here using the \"Authorize\" button."
      }
    },
    "schemas": {
...
      "DisabledCommandConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "guildId": {
            "type": "string"
          },
          "commandName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "YoutubeVideoConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "youtubeChannelId": {
            "type": "string"
          },
          "discordChannelId": {
            "type": "string"
          },
          "lastVideoId": {
            "type": "string"
          },
          "customMessage": {
            "type": "string"
          },
          "guildId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/disabledCommands": {
      "get": {
        "summary": "List all disabled commands",
        "tags": [
          "DisabledCommands"
        ],
        "responses": {
          "200": {
            "description": "List of disabled commands",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisabledCommandConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a disabled command",
        "tags": [
          "DisabledCommands"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisabledCommandConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/disabledCommands/{id}": {
      "delete": {
        "summary": "Remove a disabled command by id",
        "tags": [
          "DisabledCommands"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/disabledCommands/guild/{guildId}": {
      "get": {
        "summary": "List all disabled commands for a specific guild",
        "tags": [
          "DisabledCommands"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guildId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of disabled commands for the guild",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisabledCommandConfig"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/youtube": {
      "get": {
        "summary": "List all YouTube channel configs",
        "tags": [
          "YouTube"
        ],
        "responses": {
          "200": {
            "description": "List of YouTube configs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/YoutubeVideoConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new YouTube config",
        "tags": [
          "YouTube"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoutubeVideoConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "summary": "Upsert (add or update) a YouTube config",
        "tags": [
          "YouTube"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoutubeVideoConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/youtube/channel": {
      "get": {
        "summary": "Get a YouTube config by youtubeChannelId and discordChannelId",
        "tags": [
          "YouTube"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "youtubeChannelId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "discordChannelId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "YouTube config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YoutubeVideoConfig"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/youtube/{id}": {
      "get": {
        "summary": "Get a YouTube config by primary key",
        "tags": [
          "YouTube"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "YouTube config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YoutubeVideoConfig"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
...
}